refactor: separate chat index lifecycle and sync
